+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - `@supercharge/session`: a Supercharge session package
 - `@supercharge/http`
+    - add `ctx()` method to request and response: allows you to access the HTTP context from the related instance
     - make `Request` macroable allowing users to decorate the request with custom functions
     - add `InteractsWithState` trait to the HTTP request
     - `InteractsWithState` trait returns a state bag instead of a plain object (please check the breaking changes)